+++ install/post-install/all.sh
@@ -4,7 +4,6 @@
 source "$ARCHUP_INSTALL/post-install/boot-logo.sh"
 source "$ARCHUP_INSTALL/post-install/plymouth.sh"
 source "$ARCHUP_INSTALL/post-install/snapper.sh"
-source "$ARCHUP_INSTALL/post-install/ufw.sh"
 source "$ARCHUP_INSTALL/post-install/pacman.sh"
 source "$ARCHUP_INSTALL/post-install/post-boot-setup.sh"
 source "$ARCHUP_INSTALL/post-install/hooks.sh"